KIX2018-14117: included user create into ContactCreate API operation, included...

See merge request Softwareentwicklung/KIXng/backend/app!2047

+++ Kernel/System/User/Preferences/DB.pm
@@ -186,7 +186,7 @@ sub DeletePreferences {
     my ( $Self, %Param ) = @_;
 
     # check needed stuff
-    for (qw(UserID Key)) {
+    for (qw(UserID)) {
         if ( !$Param{$_} ) {
             $Kernel::OM->Get('Log')->Log(
                 Priority => 'error',
@@ -198,13 +198,20 @@ sub DeletePreferences {
 
     my $DBObject = $Kernel::OM->Get('DB');
 
+    my $SQL = "DELETE FROM $Self->{PreferencesTable} WHERE $Self->{PreferencesTableUserID} = ?";
+    my @Bind = (
+        \$Param{UserID}
+    );
+
+    if ( $Param{Key} ) {
+        push @Bind, \$Param{Key};
+        $SQL .= "AND $Self->{PreferencesTableKey} = ?";
+    }
+
     # delete old data
     return if !$DBObject->Do(
-        SQL => "
-            DELETE FROM $Self->{PreferencesTable}
-            WHERE $Self->{PreferencesTableUserID} = ?
-                AND $Self->{PreferencesTableKey} = ?",
-        Bind => [ \$Param{UserID}, \$Param{Key} ],
+        SQL  => $SQL,
+        Bind => \@Bind,
     );
 
     # delete cache
